added team_name to PlayerDetails. Refactored to move helper methods to helpers.py
