+++ src/data_models/DefensiveStats.py
@@ -3,7 +3,7 @@ from sqlalchemy import Column, Integer
 from sqlalchemy.sql.schema import ForeignKey
 from sqlalchemy.sql.sqltypes import Float, String
 
-from constants import Base
+from src.constants import Base
 
 
 @dataclass
