+++ src/data_models/TeamInfo.py
@@ -6,7 +6,7 @@ from sqlalchemy import(
     String
     )
 
-from constants import Base
+from src.constants import Base
 
 @dataclass
 class TeamInfo(Base):
