+++ src/models/Stats.py
@@ -1,6 +1,24 @@
 from src.models.Player import PlayerDetails
 
 
+class RushingStats:
+        def __init__(
+                self,
+                rush_att,
+                rush_yards,
+                ya_contact,
+                broke_tkls,
+                fumbles,
+                twenty_plus_yd_runs
+        ):
+                self.rush_att = rush_att
+                self.rush_yards = rush_yards
+                self.ya_contact = ya_contact
+                self.broke_tkls = broke_tkls
+                self.fumbles = fumbles
+                self.twenty_plus_yd_runs = twenty_plus_yd_runs
+
+
 class PassingStats:
     def __init__(
                 self,
@@ -22,6 +40,17 @@ class PassingStats:
         self.completions = completions
         self.ints = ints
         self.pass_att = pass_att
+
+
+
+class PlayerRushingStats:
+        def __init__(
+                self,
+                player_details: PlayerDetails,
+                rushing_stats: RushingStats
+        ):
+                self.player_details = player_details
+                self.rushing_stats = rushing_stats
 
 
 class PlayerPassingStats:
